move evaluator tests to tests/evaluators
