+++ tests/evaluators/test_ParaphraseMiningEvaluator.py
@@ -1,0 +1,24 @@
+"""
+Tests the correct computation of evaluation scores from BinaryClassificationEvaluator
+"""
+
+from __future__ import annotations
+
+from sentence_transformers import (
+    SentenceTransformer,
+    evaluation,
+)
+
+
+def test_ParaphraseMiningEvaluator(paraphrase_distilroberta_base_v1_model: SentenceTransformer) -> None:
+    """Tests that the ParaphraseMiningEvaluator can be loaded"""
+    model = paraphrase_distilroberta_base_v1_model
+    sentences = {
+        0: "Hello World",
+        1: "Hello World!",
+        2: "The cat is on the table",
+        3: "On the table the cat is",
+    }
+    data_eval = evaluation.ParaphraseMiningEvaluator(sentences, [(0, 1), (2, 3)])
+    metrics = data_eval(model)
+    assert metrics[data_eval.primary_metric] > 0.99
